Change `From` to `TryFrom`
Per the trait documentation, `From` may only be implemented for
infallible conversions. Fallible conversions MUST use `TryFrom` or other
means. The `From` implementation in the crate was breaking this contract
so this change fixes it by changing the imple to `TryFrom` and
introducing an error type reporting the error rather than panicking.