Represent non-zero amounts using `NonZeroU64`
Some amounts MUST NOT be zero and the code so far was checking that
"manually". However, Rust has a native way to express this withoug
explicit checks by using `NonZeroU64` type which also unlocks some
optimizations. Thus this commit changes the type.

There's a recomputed sum that is now handled by unwrapping. This can be
later improved together with other deserialization changes.