Change `payout_points` from `Vec` to an array
The field `payout_points` was a `Vec` despite being required to have
exactly two items. The check had to be explicit in the code, it required
extra allocation and it was harder to optimize accesses.

This simply changes it to array which works fine because of how similar
the APIs are.