Make parsing/validation idiomatic
It's unusual to construct invalid objects in Rust and call validation
functions on them. Instead all objects should be constructed as valid.
So far the code was breaking this in case of deserialization because
validation is more complicated than what `serde` natively supports.

We resolve this by using a technique I have siginificant experience
with: we duplicate the types where one represents validated data and the
other unvalidated (raw) data. Then we implement functions that return
the validated data after performing validation, proving that validation
did happen.